selenium/requirements.txt
# Selenium関連
selenium>=4.10.0
webdriver-manager>=4.0.0

# 環境変数管理
python-dotenv>=1.0.0

# HTMLパース
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ユーティリティ
requests>=2.31.0
urllib3>=2.0.0